Pull NettyTransportBase from WsTransport
